+++ assets/javascripts/kitten/components/form/checkbox.js
@@ -1,19 +1,25 @@
 import React from 'react'
 import classNames from 'classnames'
+import deprecated from 'react-prop-types/lib/deprecated'
 
 export class Checkbox extends React.Component {
   render() {
-    const { id, text, ...others } = this.props
+    const { className,
+            id,
+            children,
+            text,
+            inputClassName,
+            ...inputProps } = this.props
 
     return (
-      <div className="k-Checkbox">
+      <div className={ classNames('k-Checkbox', className) }>
         <input id={ id }
                type="checkbox"
-               className="k-Checkbox__input"
-               { ...others } />
+               className={ classNames('k-Checkbox__input', inputClassName) }
+               { ...inputProps } />
 
         <label htmlFor={ id } className="k-Checkbox__label">
-          { text }
+          { children || text }
         </label>
       </div>
     )
@@ -21,5 +27,9 @@ export class Checkbox extends React.Component {
 }
 
 Checkbox.defaultProps = {
-  text: 'Filter 1',
+  children: 'Filter 1',
+}
+
+Checkbox.propTypes = {
+  text: deprecated(React.PropTypes.string, 'Use `children` prop instead')
 }