+++ spec/dummy/client/entries/app-kitten.js
@@ -10,6 +10,7 @@ import LinkBox from 'kitten/components/box/link-box'
 import { Button } from 'kitten/components/buttons/button'
 import ButtonImageWithTextAndBadge from 'kitten/components/buttons/button-image-with-text-and-badge'
 import ButtonTooltipIcon from 'kitten/components/buttons/button-tooltip-icon'
+import { TagButton } from 'kitten/components/buttons/tag-button'
 
 // Dev
 import { DevGrid } from 'kitten/components/dev/dev-grid'
@@ -100,6 +101,7 @@ import { KarlParagraph,
 import { KarlButtonIconHeartRight,
          KarlButtonIconHeartLeft
        } from 'kitten/karl/buttons/button-heart-icon-words'
+import { KarlMakerWhoAmIStep } from 'app/maker/who-am-i/components/container'
 
 ReactOnRails.register({
   // Box
@@ -110,6 +112,7 @@ ReactOnRails.register({
   Button,
   ButtonImageWithTextAndBadge,
   ButtonTooltipIcon,
+  TagButton,
 
   // Dev
   Checkbox,
@@ -199,4 +202,5 @@ ReactOnRails.register({
   KarlParagraphTertiary,
   KarlParagraphQuaternary,
   KarlParagraphWithLink,
+  KarlMakerWhoAmIStep,
 })